The time needed to commit is getting a bit too long for my liking, takes about 2 - 3 minutes just to commit, add -all, status etc.
Anyway, in this commit, I have my concurrent map up and running,
Fortunately, I managed to be able to insert and retrieve objects. The
spinlock is currently reentrant, but does not relinquish the lock yet
because, well of course, the injected code in the compiler hasn't been
completed yet.

I'll be doing an immediate pull after this push.