Made some rather significant improvements, if I do say so myself.
The concurrent map is 10x more efficient than it was before. Now it
actually supports concurrent access among many different Goroutines. In
fact, with a thousand of Goroutines, it only came behind the normal map
(synchronized with a Mutex) by a second, but thats do to optimization of
the map itself. Currently, the real bottleneck is the amount of
collision, but I plan to fix that by increasing the size of bucket.

Also, as I mentioned above, it supports concurrent access; this means
that it needs to release the spinlock after the indirection of the
returned pointer (value), and so I needed to add that as a hook in the
compiler.