I decided to move from a static array to a dynamic one. Due to the fact that Go does not natively support dynamic arrays (Slices do not work unfortunately for me), I had to do some runtime magic to get that going (see mallocgc).
Also, some good news is that my map apparently is 2 - 3X as fast as the
normal Hashmap... bad news is that it doesn't actually work yet. I can
definitely say that the speed would be the same as it repeatedly
attempts to write to it, although it writes the wrong values (Gotta
double check how hashmap.go does it). Hence, once it is stable, I can
foresee it being a very high performant concurrent map, and I can't
wait. Still 4 weeks left, so not too much of a huge deal.