The concurrent hashmap, as benchmarked, has some really bad bottlenecks and there are some issues (I.E Fast Types) that need to be resolved, but overall, it is rather fast.
When I benchmarked for 1000 Goroutines to insert 1000 items
concurrently, it was 2 - 3 times faster, consistently. However, when I
moved the amount of Goroutines to 2000, effectively doubling the number
of elements from one million to two million, it lagged behind, becoming
3 times slower on average, and it is due to the amount of time spent
converting and creating a new array. From now on, I will not allocate
all nodes ahead of time... this time instead I will just allocate it as
I would a normal linked list. This way I can also do a move of a linked
node to another bucket cell.