Developed a benchmark with reliable results.
For one, as Dr. Spears has confirmed, the concurrent map is highly scalable. While with only one Goroutine acessing the map, we see that our map takes twice as long as the original.
However, at a more common 4 to 8 Goroutines (or rather, Goroutines matching the GOMAXPROCS variable) are 2 - 3 times faster than the normal map (protected by a mutex). Reader-Writer
locks would take even longer, I'm certain. The benchmark isn't really extensive yet, however it does show promise for the map at hand.

Currently, Insertion, Retrieval, and Deletion all work and are fully implemented. They leave room for optimization, but even so, with a standard GOMAXPROCS rate of 4 to 8 (meaning, on
most machines, this will be the case), it shows a LOT of promise. Iteration is up next.