Currently, I have the map's iterator working, but it isn't as fast as I'd like, although I'm not sure how much more that can be helped.
The map seems to take, up to 5 - 10x longer to traverse than the a
synchronized or reader-writer one, but then again, those protect the
entire for loop.

The only reason mine is so scalable and fast is the reason that the
locks are not held for long at all. The iterator currently only holds 8
elements at a time, and hence can process 8 elements at once; when it
needs more, it will grab the lock on the next bucket, do a typememmove
(which misleadingly only does a copy), and release the lock immediately
after. This does not slow down insertions, retrievals or deletions,
which is why those operations are so fast.

Truth be told, I don't think it's that big of a deal that it takes so
long: the amount of time is about equivalent to the time it takes for a
normal synchronized or reader-writer map to insert or delete. Hence I
think it is a good enough trade-off.