1) Began the attempt to implement a new keyword to Go, which should not result in any collision...
runtime.Interlocked

This keyword will have one or two effects on the user's code.

First being that runtime.Interlocked, if in the scope of a for loop
range block, as such...

for k, v := range runtime.Interlocked m { ... }

Then the key and value will be guaranteed to exist in the map, and the
value will be returned by pointer, allowing changes to it inside the
map. Note that the bucketData holding it will be locked for the entire
duration, hence these kind of for loops should be short. The normal
syntax for a map is supported, and like the default hashmap, it returns
them by value (and the bucket will give up it's lock).

Secondly, in the context of it's own statement

runtime.Interlocked m { ... }

Any keys attempted to be used will be evaluated, and grabbed ahead of
time. This allows a small "critical section".

2) The iterator will now be randomized (currently in progress).

Currently, I believe (and Michael Spears has confirmed) that the issue
of starting at the beginning causes a lot of convoying and bottlenecks.
Instead, I will have the iterator start at a cell in a randomized and
nested bucketArray based on a randomly generated seed. This allows
greater concurrency, as the chances of multiple iterators contending and
convoying over a cell is very unlikely in deeply nested and filled out
maps. In maps which have a limited of elements, then having potentially
16 concurrent iterators isn't too bad either way.