1) Added randomized ordering for the concurrent iterator. Unfortunately, unlike the map, it needs to be able to traverse multiple (arbitrarily many) nestings, and gets complicated fast once you have a real scenario. For example, if a bucketHdr's bucket points to nil, it's empty and we need to move on to the next. If we wrap all the way around, we must go back one (keeping track of that as well). If we find a recursive bucketHdr pointing to a bucketArray, we must store that information so we can go back. This information must also be usable by the cmapiternext function, so we effectively end up repeating portions of code due to it's complexity. I also use goto statements, which will be removed if possible in the next commit.
The randomized ordering does help performance of the iterator, in a way,
and hurts in another. The previous simplistic model faired well under
concurrency as it went sequentially, but was subject to the bottlenecks
of contention from the lock convoying (I.E, an iterator is waiting in
line for another iterator to finish, since they all started at the same
index). It helps in the fact that when you do not have much contention,
the only contention is from waiting on other iterators. If the map is
under heavy contention, then convoying with other iterators is the least
of the worries. The complexity of having to find a random recursive
nested bucketData to begin at, and then save and restore the state to be
able to continue, has it's overhead in and of itself. It now takes about
2x as long as before when under contention, but 1/3 the time when under
no contention. A victory? Maybe, but bittersweet.

2) I improved the granularity and efficiency of the busy-wait algorithm
by going from a Test-And-Set strategy to a Test-And-Test-And-Set
strategy. I found out today that RMW atomic operations, I.E
Compare-And-Swap, cause a lot of overhead with cache invalidation and
"bus locking" to ensure atomicity. Now I do a simple Load, check for
certain flags (and invariants which or relaxed assumptions, I.E if the
hdr.bucket == nil check for the iterator) and act on that, which
includes the Test-And-Test-And-Set for the spin lock.

This actually has marvelously increased the performance of my
application. The next thing I can think of is adding exponential backoff
(gotta figure out how to cause Goroutines to sleep without stalling
processor in runtime) or ticket locks (which would also need exponential
backoff).