1) Created a new benchmark which will simualate a real-life scenario for it's usage. It is a basic, but more realistic one, that has simulates having multiple readers doing work, producing data, and adding it to a concurrent map (8 at once). You have another Goroutine constantly iterating it, which spawns helper threads to process said data. Each helper Goroutine will be processing that data and adding them to another concurrent map.
Normally, one would have one Goroutine read it, pass it through a
channel, have a thread iterate and spawn a Goroutine to process it,
everything protected by a mutex. The real issue here is the lack of
concurrency. Yes you ahve concurrent processors, but even that is
limited. Nevermind, I'm rambling at this point...

2) Attempted to begin implementing more of sync.Interlocked.