Finally managed a way to acquire the keys... in a very ugly, hackish fashion. Apparently, the Golang developers did not have it's mkcall* functions check if it was variadic, and hence it only took the first key, instead of all keys passed, since of course when they designed it, they didn't need it (Go figure).
Anyway, I can confirm that the keys are now being passed to it, and it
currently makes a temporary copy of each key and passes it to
mapacquire. Tomorrow, I will add the VARKILL to free up those temporary
slots as well, and also implement maprelease, as well as an actual
algorithm that allows acquisition of multiple keys.