Managed to create the symbol passed, as well as handle assigning to it as well. Hence...
sync.Interlocked obj := map[key] {...}

'obj' would be the value holding the value found at 'map[key]'. This may seem
relatively ugly, but overall, it's actually as close as I'll get to what I wish to
achieve.

Currently, it holds the lock for the entire duration, but it does not release the
lock, yet. This should be a relatively easy addition.

The goal is to have

sync.Interlocked n := m[k] {
	n++
}

Be a reality. That is, all operations to 'n' will be atomic, as the bucket held at
m[k] is locked until it finishes it's block. Why is this important? Increments, like
that, are guaranteed to be atomic and accurate. While 'n' is a copy, it will
automatically update the value inside of the map on-demand.