Made quite a few changes. I forget the precise changes I've made since last commit, but here are a few.
1) sync.Interlocked range loops currently work, and have efficient
randomized iteration start positions. As a reminder, sync.Interlocked
grabs the lock on a bucket, and holds on to it for the entire duration
of the iteration over it, rather than (like the default range) grabs the
lock, makes an atomic snapshot, and then releases. It is a bit faster
than the default, but not by much.

2) Changed the default number of buckets from 16 to 32. Resulted in a
phenemonal increase in overall performance for the iteration (an
astounding 2x).

3) Added exponential backoff with yielding for lower contention.
Basically, after a failed spin to check if the lock has been released,
it will yield(). If it yields enough times and the lock still hasn't
been released (from the perspective of the contending Goroutine), it
will backoff for one microsecond, exponentially increasing up to 33ms.

4) sync.Interlocked region, 'sync.Interlocked map[key]', is working
perfectly.