After discovering that tools like 'go test' use 'src/Go' to compile the test and benchmark functions, and that the new keyword 'sync.Interlocked' would need to also be implemented there, I decided to just do it the old fashioned way.
Now, the benchmark for intset is complete, and the benchmark outputs the
appropriate information in a way that can be put into a .csv file, and
from there I can create graphs. As well, I now have all sources
indviidualized to each Goroutine, as the default global source, of
course, is synchronized, and this resulted in gains as high as 4 - 5x.
Amazing.